separated server logic from main
